dry-run option for commit and rollback commands

+++ src/Commands/RollbackCommand.php
@@ -15,6 +15,7 @@ class RollbackCommand extends Command
 
 	private const ARGUMENT_LIMIT = 'limit';
 	private const OPTION_STRATEGY = 'strategy';
+	private const OPTION_DRY_RUN = 'dry-run';
 
 	protected function configure()
 	{
@@ -26,7 +27,8 @@ class RollbackCommand extends Command
 				InputArgument::OPTIONAL,
 				'Rollback strategy (by commit "' . Kernel::ROLLBACK_BY_DATE . '" or by migration "' . Kernel::ROLLBACK_BY_ORDER . '")',
 				Kernel::ROLLBACK_BY_DATE
-			);
+			)
+			->addOption(self::OPTION_DRY_RUN, 'd', InputArgument::REQUIRED, 'Run in dry-run mode');
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output): int
@@ -38,6 +40,11 @@ class RollbackCommand extends Command
 		$table = $config->getTable();
 		$migrationsLimit = $input->getArgument(self::ARGUMENT_LIMIT);
 		$rollbackStrategy = $input->getOption(self::OPTION_STRATEGY);
+		$isDryRun = $input->getOption(self::OPTION_DRY_RUN);
+
+		if ($isDryRun) {
+			$this->writelnFormatted(' DRY-RUN ', 'black', 'cyan');
+		}
 
 		if (!in_array($rollbackStrategy, [
 			Kernel::ROLLBACK_BY_ORDER,
@@ -46,18 +53,21 @@ class RollbackCommand extends Command
 			throw new Exception('Unknown rollback strategy: ' . $rollbackStrategy);
 		}
 
-		$this->writelnCyan('Rollback strategy: ' . $rollbackStrategy);
+		$this->writelnCyan('Using rollback strategy: by ' . $rollbackStrategy);
 
 		$rollbackPerformed = FALSE;
 		$count = 0;
-		while ($lastMigration = $this->kernel->getLastMigration($rollbackStrategy)) {
+		$migrations = $this->kernel->getAllMigrations($rollbackStrategy);
+		foreach ($migrations as $migration) {
 			$rollbackPerformed = TRUE;
-			$migrationFile = $lastMigration[$table->fileName];
+			$migrationFile = $migration[$table->fileName];
 			require_once $this->kernel->getMigrationPath($migrationFile);
 
-			$this->write('Migration "' . $migrationFile . '" rollback ... ');
 			/** @var Migration $migrationClass */
 			$migrationClass = $this->kernel->parseMigrationClassName($migrationFile);
+
+			$this->write('Migration "' . $migrationFile . '" rollback ... ');
+			$count++;
 
 			$connection->begin();
 			try {
@@ -65,16 +75,17 @@ class RollbackCommand extends Command
 					throw new Exception('Migration is breakpoint and thus rollback is unable');
 				}
 
-				$connection->nativeQuery($migrationClass::down());
+				if (!$isDryRun) {
+					$connection->nativeQuery($migrationClass::down());
 
-				$connection->delete($table->getName())
-					->where('%n = %i', $table->primaryKey, $lastMigration[$table->primaryKey])
-					->execute();
+					$connection->delete($table->getName())
+						->where('%n = %i', $table->primaryKey, $migration[$table->primaryKey])
+						->execute();
+				}
 
 				$connection->commit();
 
 				$this->writelnSuccess(' DONE ');
-				$count++;
 			} catch (Exception $e) {
 				$connection->rollback();
 				$this->writelnError(' FAILURE ');
